+++ server/utils/imageHelpers.js
